# MRG, ENH: Speed up clustering using numba