--- a/mne/stats/tests/test_cluster_level.py
+++ b/mne/stats/tests/test_cluster_level.py
@@ -9,7 +9,7 @@ import os
 import numpy as np
 from scipy import sparse, linalg, stats
 from numpy.testing import (assert_equal, assert_array_equal,
-                           assert_array_almost_equal)
+                           assert_array_almost_equal, assert_allclose)
 import pytest
 
 from mne.parallel import _force_serial
@@ -18,7 +18,7 @@ from mne.stats.cluster_level import (permutation_cluster_test, f_oneway,
                                      spatio_temporal_cluster_test,
                                      spatio_temporal_cluster_1samp_test,
                                      ttest_1samp_no_p, summarize_clusters_stc)
-from mne.utils import run_tests_if_main, _TempDir, catch_logging, check_version
+from mne.utils import run_tests_if_main, catch_logging, check_version
 
 
 n_space = 50
@@ -58,11 +58,11 @@ def test_thresholds():
     with catch_logging() as log:
         with pytest.warns(RuntimeWarning, match='threshold is only valid'):
             out = permutation_cluster_1samp_test(X, stat_fun=my_fun,
-                                                 verbose=True)
+                                                 seed=0, verbose=True)
     log = log.getvalue()
     assert str(want_thresh)[:6] in log
     assert len(out[1]) == 1  # 1 cluster
-    assert 0.03 < out[2] < 0.05
+    assert_allclose(out[2], 0.033203, atol=1e-6)
     # between subjects
     Y = rng.randn(10, 1, 1)
     Z = rng.randn(10, 1, 1) - 0.7
@@ -74,18 +74,18 @@ def test_thresholds():
     with catch_logging() as log:
         with pytest.warns(RuntimeWarning, match='threshold is only valid'):
             out = permutation_cluster_test(X, tail=1, stat_fun=my_fun,
-                                           verbose=True)
+                                           seed=0, verbose=True)
     log = log.getvalue()
     assert str(want_thresh)[:6] in log
     assert len(out[1]) == 1  # 1 cluster
-    assert 0.03 < out[2] < 0.05
+    assert_allclose(out[2], 0.041992, atol=1e-6)
     with pytest.warns(RuntimeWarning, match='Ignoring argument "tail"'):
         permutation_cluster_test(X, tail=0)
 
 
-def test_cache_dir():
+def test_cache_dir(tmpdir):
     """Test use of cache dir."""
-    tempdir = _TempDir()
+    tempdir = str(tmpdir)
     orig_dir = os.getenv('MNE_CACHE_DIR', None)
     orig_size = os.getenv('MNE_MEMMAP_MIN_SIZE', None)
     rng = np.random.RandomState(0)
@@ -157,11 +157,15 @@ def test_permutation_step_down_p():
         permutation_cluster_1samp_test(X, threshold=thresh,
                                        step_down_p=0.0)
     assert_equal(np.sum(p_old < 0.05), 1)  # just spatial cluster
+    p_min = np.min(p_old)
+    assert_allclose(p_min, 0.003906, atol=1e-6)
     t, clusters, p_new, H0 = \
         permutation_cluster_1samp_test(X, threshold=thresh,
                                        step_down_p=0.05)
     assert_equal(np.sum(p_new < 0.05), 2)  # time one rescued
     assert np.all(p_old >= p_new)
+    p_next = p_new[(p_new > 0.004) & (p_new < 0.05)][0]
+    assert_allclose(p_next, 0.015625, atol=1e-6)
 
 
 def test_cluster_permutation_test():
@@ -173,12 +177,9 @@ def test_cluster_permutation_test():
         T_obs, clusters, cluster_p_values, hist = permutation_cluster_test(
             [condition1, condition2], n_permutations=100, tail=1, seed=1,
             buffer_size=None)
+        p_min = np.min(cluster_p_values)
         assert_equal(np.sum(cluster_p_values < 0.05), 1)
-
-        T_obs, clusters, cluster_p_values, hist = permutation_cluster_test(
-            [condition1, condition2], n_permutations=100, tail=1, seed=1,
-            buffer_size=None)
-        assert_equal(np.sum(cluster_p_values < 0.05), 1)
+        assert_allclose(p_min, 0.01, atol=1e-6)
 
         # test with 2 jobs and buffer_size enabled
         buffer_size = condition1.shape[1] // 10
@@ -206,13 +207,16 @@ def test_cluster_permutation_t_test():
                  partial(ttest_1samp_no_p, sigma=1e-1)]
 
     for stat_fun in stat_funs:
-        for condition1 in (condition1_1d, condition1_2d):
+        for condition1, p in ((condition1_1d, 0.01),
+                              (condition1_2d, 0.01)):
             # these are so significant we can get away with fewer perms
             T_obs, clusters, cluster_p_values, hist =\
                 permutation_cluster_1samp_test(condition1, n_permutations=100,
                                                tail=0, seed=1,
                                                buffer_size=None)
             assert_equal(np.sum(cluster_p_values < 0.05), 1)
+            p_min = np.min(cluster_p_values)
+            assert_allclose(p_min, p, atol=1e-6)
 
             T_obs_pos, c_1, cluster_p_values_pos, _ =\
                 permutation_cluster_1samp_test(condition1, n_permutations=100,
@@ -428,7 +432,7 @@ def test_permutation_connectivity_equiv():
         t, clusters, p, H0 = \
             permutation_cluster_1samp_test(
                 X, threshold=thresh, connectivity=conn, n_jobs=2,
-                max_step=max_step, stat_fun=stat_fun)
+                max_step=max_step, stat_fun=stat_fun, seed=0)
         # make sure our output datatype is correct
         assert isinstance(clusters[0], np.ndarray)
         assert clusters[0].dtype == bool
@@ -438,6 +442,7 @@ def test_permutation_connectivity_equiv():
         # should come up empty
         inds = np.where(p < 0.05)[0]
         assert_equal(len(inds), count)
+        assert_allclose(p[inds], 0.03125, atol=1e-6)
         if isinstance(thresh, dict):
             assert_equal(len(clusters), n_time * n_space)
             assert np.all(H0 != 0)