--- a/tutorials/stats-sensor-space/plot_stats_cluster_1samp_test_time_frequency.py
+++ b/tutorials/stats-sensor-space/plot_stats_cluster_1samp_test_time_frequency.py
@@ -8,7 +8,7 @@ in time-frequency power estimates. It uses a non-parametric
 statistical procedure based on permutations and cluster
 level statistics.
 
-The procedure consists in:
+The procedure consists of:
 
   - extracting epochs
   - compute single trial power estimates