--- a/tutorials/stats-source-space/plot_stats_cluster_spatio_temporal_2samp.py
+++ b/tutorials/stats-source-space/plot_stats_cluster_spatio_temporal_2samp.py
@@ -8,7 +8,6 @@ Tests if the source space data are significantly different between
 The multiple comparisons problem is addressed with a cluster-level
 permutation test across space and time.
 """
-
 # Authors: Alexandre Gramfort <alexandre.gramfort@telecom-paristech.fr>
 #          Eric Larson <larson.eric.d@gmail.com>
 # License: BSD (3-clause)
@@ -111,4 +110,3 @@ brain = stc_all_cluster_vis.plot('fsaverage', hemi='both',
                                  views='lateral', subjects_dir=subjects_dir,
                                  time_label='Duration significant (ms)',
                                  clim=dict(kind='value', lims=[0, 1, 40]))
-brain.save_image('clusters.png')