--- a/tutorials/stats-source-space/plot_stats_cluster_time_frequency_repeated_measures_anova.py
+++ b/tutorials/stats-source-space/plot_stats_cluster_time_frequency_repeated_measures_anova.py
@@ -237,4 +237,4 @@ plt.show()
 
 ###############################################################################
 # Both cluster level and FDR correction help get rid of
-# putatively spots we saw in the naive f-images.
+# potential spots we saw in the naive f-images.