This code snippet:
```
import time
import numpy as np
import mne
src_fname = (mne.datasets.sample.data_path() +
             '/subjects/fsaverage/bem/fsaverage-ico-5-src.fif')
X = np.random.RandomState(0).randn(8, 5, 20484) * 0.1
src = mne.read_source_spaces(src_fname)
connectivity = mne.spatial_src_connectivity(src)
t0 = time.time()
mne.stats.spatio_temporal_cluster_1samp_test(
    X, connectivity=connectivity, n_jobs=1, buffer_size=None, verbose=True,
    max_step=1)
print(time.time() - t0)
```
on `master` takes 16.7 sec, on this PR it takes 4.5 sec.

Changing `max_step=2` then on `master` it takes 23.1 sec, on this PR it takes 6.1 sec.

So `numba` gets us a ~4x speedup in time-consuming code. ~~I don't love the code dup, but both paths should be tested for correctness by Travis.~~ On systems with numba, both code paths are tested.

The changes to tests were developed on and pass on `master`, just wanted to add some extra sanity checks about the p values. Touched a bunch of clustering examples so that I can check the outputs on CircleCI, but they should match.