@larsoner 2019-08-13
Pushed a commit that makes it so that on machines that have `numba`, both `numba` and non-`numba` code paths are tested. I think this is a sustainable way to have `numba`-conditional branching in our code.