@larsoner 2019-08-14
I checked the CircleCI outputs and the figures all looked the same.