@agramfort 2019-08-16
thx @larsoner